requirements.txt
Pillow == 6.2.1
boto3 == 1.10.6
selenium == 3.141.0